backend/requirements.txt
Flask==2.3.0
firebase-admin==6.0.1
python-dotenv==1.0.0
APScheduler==3.10.1
sendgrid==6.11.0
twilio==8.6.0
requests==2.31.0
gunicorn==20.1.0
